# contracts/openapi/intelligence-api.yaml
openapi: 3.1.0
info:
  title: Planetary Intelligence API
  version: 0.1.0
servers:
  - url: https://api.example.com
paths:
  /v1/briefs/now:
    post:
      summary: Generate a 30-second intelligence brief
      operationId: generateNowBrief
      tags: [Briefs]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NowBriefRequest'
      responses:
        '200':
          description: Intelligence brief
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NowBriefResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '429':
          description: Rate limited
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NowBriefRequest:
      type: object
      required: [tenantId, userId, scope]
      properties:
        tenantId:
          type: string
        userId:
          type: string
        scope:
          type: object
          properties:
            regions:
              type: array
              items:
                type: string
            entities:
              type: array
              items:
                type: string
            domains:
              type: array
              items:
                type: string
    Recommendation:
      type: object
      required: [action, rationale, urgency]
      properties:
        action:
          type: string
        rationale:
          type: string
        urgency:
          type: string
          enum: [low, medium, high, critical]
    EvidenceRef:
      type: object
      required: [sourceId, claimId]
      properties:
        sourceId:
          type: string
        claimId:
          type: string
    NowBriefResponse:
      type: object
      required:
        [
          generatedAt,
          changed,
          whyItMatters,
          likelyNext,
          recommendedActions,
          confidence,
          uncertainty,
          evidence,
        ]
      properties:
        generatedAt:
          type: string
          format: date-time
        changed:
          type: array
          items:
            type: string
        whyItMatters:
          type: array
          items:
            type: string
        likelyNext:
          type: array
          items:
            type: string
        recommendedActions:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
        confidence:
          type: number
          minimum: 0
          maximum: 1
        uncertainty:
          type: object
          required: [score, factors]
          properties:
            score:
              type: number
              minimum: 0
              maximum: 1
            factors:
              type: array
              items:
                type: string
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceRef'